feat: add region filter
This commit adds a regional filter to the python script.  This will
take in the following environment variables:

INCLUDE_REGIONS = regions to be included in script (defaults to all)
EXCLUDE_REGIONS = regions to be excluded in script (defaults to none)

If both are included, the difference of the 2 lists will be checked
against.

Signed-off-by: Dustin Scott <dscott@redhat.com>